FIXED DYNAMIC ARRAYS AND POINTERS

+++ src/main.cc
@@ -41,6 +41,58 @@ int main(int argc, char** argv) {
     func_dir = "src/functions";
 #endif
 
+#ifdef __linux__
+    m_preprocessor.insert("__LINUX__");
+#elif __WIN32
+      m_preprocessor.insert("__WIN__");
+      #ifdef __WIN64
+        m_preprocessor.insert("__WIN64__");
+      #endif
+#elif __APPLE__
+      m_preprocessor.insert("__APPLE__");
+      #ifdef TARGET_OS_MAC
+        m_preprocessor.insert("__MACOS__");
+      #elif TARGET_OS_IPHONE("__IOS__");
+      #endif
+#endif
+
+#if defined(__i386__)
+    m_preprocessor.insert("__i386__");
+#elif defined(__x86_64__) || defined(_M_X64)
+  m_preprocessor.insert("__X86_64__");
+#endif
+
+#if defined(__arm__)
+    m_preprocessor.insert("__arm__");
+#elif defined(__aarch64__)
+    m_preprocessor.insert("__aarch64__");
+#endif
+
+
+#if defined(__ppc64__) || defined(_M_PPC)
+    m_preprocessor.insert("__ppc64__");
+#elif defined(__riscv)
+    m_preprocessor.insert("_M_PPC");
+    m_preprocessor.insert("__RICV_V__");
+#endif
+
+
+#if INTPTR_MAX == INT64_MAX
+    m_preprocessor.insert("__64BYTES__");
+#elif INTPTR_MAX == INT32_MAX
+   m_preprocessor.insert("__32BYTES__"); 
+#else
+    #error "Arquitectura desconocida"
+#endif
+
+#if __BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__
+    m_preprocessor.insert("__ORDER_LITTLE_ENDIAN__");
+#elif __BYTE_ORDER__ == __ORDER_BIG_ENDIAN__
+    m_preprocessor.insert("__ORDER_BIG_ENDIAN__");
+#endif
+
+m_preprocessor.insert("__YASOS_ORIGINAL_COMPILER__");
+
     const std::string cache_directory = "/home/" + username +  "/.cache/yasos/";
     if (system(("test -d " + cache_directory).c_str()) != 0) {
         if (system(("mkdir " + cache_directory).c_str()) != 0) {
@@ -49,12 +101,13 @@ int main(int argc, char** argv) {
     }
 
     bool keep_asm = false;
+    bool keep_ll = false;
     bool generate_asm_only = false;  // -S
     bool compile_only = false;       // -c
     bool generate_shared = false;    // -shared
     std::string output_file = "out"; // -o
     bool generate_executable = false;
-	bool print_link_command = false;
+	  bool print_link_command = false;
     std::vector<std::string> input_files;
 
     init_debug();
@@ -73,6 +126,8 @@ int main(int argc, char** argv) {
             generate_shared = true;
         } else if (arg == "--preserve-asm") {
             keep_asm = true;
+        } else if (arg == "--preserve-ll") {
+          keep_ll = true;
         } else if (arg == "--print-link-command"){
 			print_link_command = true;
 		} else {
@@ -133,7 +188,7 @@ int main(int argc, char** argv) {
         std::string gen_name = base_name;
 
         Generator generator(program.value(), gen_name, std::move(TheModule));
-        std::string asm_code = generator.gen_prog();
+        generator.gen_prog();
         std::string ll_file = base_name + ".ll";
         std::error_code EC;
         llvm::raw_fd_ostream ll_out(ll_file, EC);
@@ -141,7 +196,7 @@ int main(int argc, char** argv) {
         ll_out.close();
 
         std::string asm_file = base_name + ".s";
-        if (!generate_asm_only) asm_file = cache_directory +  base_name + ".s";
+        if (!generate_asm_only) asm_file = base_name + ".s";
         std::string obj_file = base_name + ".o"; 
 
         std::string llc_cmd = "llc -relocation-model=pic -filetype=asm " + ll_file + " -o " + asm_file;
@@ -164,8 +219,10 @@ int main(int argc, char** argv) {
             object_files.push_back(obj_file);
         }
 
-        const std::string command_rm_ll = "rm " + ll_file;
-        system(command_rm_ll.c_str());
+        if (!keep_ll) {
+          const std::string command_rm_ll = "rm " + ll_file;
+          system(command_rm_ll.c_str());
+        }
 
         if (generate_asm_only) {
             continue;
@@ -195,9 +252,9 @@ int main(int argc, char** argv) {
     std::string link_command;
 
     if (generate_executable) {
-        link_command = "g++ -no-pie -o " + output_file + " ";
+        link_command = "g++ -g -no-pie -o " + output_file + " ";
     } else {
-        link_command = "g++ -no-pie ";
+        link_command = "g++ -g -no-pie ";
     }
 
     if (generate_shared) {
